no need for $data
no need for $data - see http://answers.semanticweb.com/questions/28985/allvaluesfrom-without-somevaluesfrom : "But when you say "IN CASE they are there", so you want to say "in case the value of the property is written in the data" or "we know that there is a value for the property"? In OWL, there is no way to restrict what is written in the data. You can only define what is true in the world, whether or not it is written."